separate splitting expression into array part of code from parseExpression into own splitExpressionIntoArray.js file and function
